requirements.txt
numpy~=1.23.5
Flask~=2.3.2
tensorflow~=2.12.0
wandb~=0.15.4
scikit-learn~=1.2.2
keras~=2.12.0
Pillow~=9.0.0